feat : Implement adding zero coffee log & completing today's coffee log
