+++ src/screens/monthly-record/monthly-record.tsx
@@ -1,7 +1,11 @@
 import React, { ReactElement, useState } from "react";
 import { View, Button } from "react-native";
 import { firebaseApp } from "@services/firebaseApp";
-import { addNormalCupRecord, addZeroCupRecord } from "@services/coffeeLog-service";
+import {
+  addNormalCupRecord,
+  addZeroCupRecord,
+  completeTodayRecord
+} from "@services/coffeeLog-service";
 import { useState as HSUseState } from "@hookstate/core";
 import { globalUserState } from "@stores/stores";
 import { Calendar, RecordList } from "@Components";
@@ -29,6 +33,7 @@ export default function MonthlyRecord({ navigation }: NavigationProps): ReactEle
   const handleButtonComplete = async () => {
     const timestamp = new Date();
     const [response, error] = await addNormalCupRecord(userEmail, shot, base, option, timestamp);
+    console.log(userEmail);
     if (error) {
       console.error(error);
     } else {
@@ -36,14 +41,25 @@ export default function MonthlyRecord({ navigation }: NavigationProps): ReactEle
     }
   };
 
-  const handleButtonZeroComplete = () => {
+  const handleButtonZeroComplete = async () => {
+    console.log(userEmail);
     const timestamp = new Date();
-    addZeroCupRecord(userEmail, timestamp);
+    const [response, error] = await addZeroCupRecord(userEmail, timestamp);
+    if (error) {
+      console.error(error);
+    } else {
+      console.log(response, "Adding coffee log succeed");
+    }
   };
 
-  const handleButtonRecordDone = () => {
+  const handleButtonRecordDone = async () => {
     const timestamp = new Date();
-    //
+    const [response, error] = await completeTodayRecord(userEmail, timestamp);
+    if (error) {
+      console.error(error);
+    } else {
+      console.log(response, "Adding coffee log succeed");
+    }
   };
 
   return (
@@ -55,11 +71,15 @@ export default function MonthlyRecord({ navigation }: NavigationProps): ReactEle
       <Button title="baseMilk" onPress={() => setBase("milk")} />
       <Button title="optionCream" onPress={() => setOption([...option, "cream"])} />
       <Button title="optionSyrup" onPress={() => setOption([...option, "syrup"])} />
-      <Button title="complete" onPress={handleButtonComplete} />
-
-      <Button title="0잔 기록" onPress={handleButtonZeroComplete} />
-
-      <Button title="오늘 마감" onPress={handleButtonRecordDone} />
+      <View style={{ marginTop: 10 }}>
+        <Button title="complete" onPress={handleButtonComplete} />
+      </View>
+      <View style={{ marginTop: 10 }}>
+        <Button title="0잔 기록" onPress={handleButtonZeroComplete} />
+      </View>
+      <View style={{ marginTop: 10 }}>
+        <Button title="오늘 마감" onPress={handleButtonRecordDone} />
+      </View>
 
       {/* <Calendar />
       <View style={styles.recordBackground}>
